// features/kits/src/utils/fetch/types.ts
export interface OptionConfig {
    commonParams?: Record<string, unknown>;
    // ssoLoginQuery?: SSOLoginQuery;
    timeout?: number;
    isRawRes?: boolean;
    unbindRolePermission?: boolean;
    /*
      useCommonErrorHandler: whether to handle error in fetch (default value is true).
      -false：handled by caller
      -true: handled in fetch
    */
    useCommonErrorHandler?: boolean;
    verify?: (T: any) => boolean;
  }